backend/requirements.txt
fastapi
uvicorn
python-dotenv
alembic
sqlalchemy
psycopg2-binary
asyncpg            # Для асинхронного PostgreSQL
aioredis             # Асинхронный Redis
celery[redis]        # Celery + Redis
aiogram             # Telegram бот
python-jose[cryptography]  # JWT токены
passlib[bcrypt]      # Хеширование паролей